% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotRainbow.R
\name{barplot_rainbow}
\alias{barplot_rainbow}
\title{Rainbow Barplot}
\usage{
barplot_rainbow(data_vector)
}
\arguments{
\item{data_vector}{data vector}
}
\value{
bar plot
}
\description{
Rainbow Barplot
}
\examples{
data <- c(10, 20, 30, 40, 50)
barplot_rainbow(data)
}
