% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanclt.R
\name{meanclt}
\alias{meanclt}
\title{Central Limit Theorem CLT Function for Sample Means}
\usage{
meanclt(n, iter)
}
\arguments{
\item{n}{Number of random variables to generate per iter.}

\item{iter}{Number of iterations to perform.}
}
\value{
A vector containing sample means.
}
\description{
This function generates random numbers from a uniform distribution,
calculates the mean of each iteration, and plots a histogram of the sample means.
}
\examples{
meanclt(n = 10, iter = 10000)
}
