% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Iterations for bootstrap sampling, 10000 by default.}

\item{x}{Vector of data.}

\item{fun}{The function to compute; default is "mean".}

\item{alpha}{Confidence level; default is 0.05.}

\item{cx}{Expansion factor for confidence interval text; default is 1.5.}

\item{...}{Additional arguments.}
}
\value{
A list containing information about the function.
}
\description{
Compute bootstrap confidence interval
}
\details{
Computes a bootstrap confidence interval for a given statistic applied to the data.
}
\examples{
x <- rnorm(100)

}
