% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Maximum Likelihood Estimation}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{The likelihood function, logbin2 by default.}

\item{theta}{Parameter values.}
}
\value{
Param value that maximizes the likelihood function.
}
\description{
Finds the parameter value that maximizes the likelihood function.
}
\examples{
logbin2 <- function(theta) {
    log(dbinom(3, prob = theta, size = 6)) +
    log(dbinom(5, prob = theta, size = 10))
}
mymaxlikg(theta = seq(0, 1, length = 10000), lfun = logbin2)
}
