% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of normal distribution.}

\item{sigma}{Standard Deviation of normal distribution.}

\item{a}{Limit for probability calculation, which will be <= a.}
}
\value{
A list containing mean, standard deviation, and the probability area.
}
\description{
Normal Curve Function
}
\examples{
myncurve(5, 10, 3)
}
