% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Find the number of tickets to sell to avoid overbooking, with discrete and continuous distributions.}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats on flight, 200 by default.}

\item{gamma}{Probability of overbooking 0.02 by default.}

\item{p}{Probability a passenger will show, 0.95 by default.}
}
\value{
A list containing: The number of tickets to be sold calculated using the discrete and continuous distribution, along with the parameters passed in the function call.
}
\description{
This function calculates the number of tickets to be sold given N, gamma, and p using discrete and continuous distributions.
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
