% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randsample.R
\name{randsample}
\alias{randsample}
\title{Make Barplots for random samples}
\usage{
randsample(n, iter = 5)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations, which is 5 by default}
}
\value{
Display barplots of created factor based on sample
}
\description{
Make random samples of size n between 1:10 with replacement, and make factor based on frequency of each number in the sample. Then, make barplot based on this factor.
}
\examples{
randsample(20, iter = 3)
randsample(10)
}
